Split the undo tests out of the factory spec
strategy-game-factory.spec.tsx covered everything the engine does in one
1062-line file, so finding the tests for any one concern meant scrolling past
nine others. undo is the most self-contained block and the second largest, so
it goes first: 834 lines stay, 194 move.

The shared fixtures move to spec-helpers.tsx — the trivial one-move game every
spec of the factory renders. The name keeps it out of vitest's *.spec.* glob.
The Headless UI warm-up becomes warmUpPlayerNameSetup() so each file that
switches to vsHuman can pay it once.

Test count is unchanged at 58 across the two files. Since the suite runs with
isolate: false, the two now share a worker context, so this was also checked
with the files run together five times, each alone, and the whole suite under
--isolate.
